+++ socks.c
@@ -81,9 +81,9 @@ int socks4_request(int socket, uint8_t cmd, struct sockaddr_in *daddr, char *use
 
     printl(LOG_VERB, "SOCKS4 reply: [%d][%d], Bytes [%d]", rep.nul, rep.status, rcount);
 
-    if (rep.nul != 0) {
-        printl(LOG_CRIT, "SOCKS4 server speaks unsupported protocol version: [%d]", rep.nul);
-        mexit(1, pfile_name);
+    if (rep.nul != 0) {                                             /* Reply SOCKS4 Request rejected or failed */
+        printl(LOG_CRIT, "SOCKS4 server speaks unsupported protocol v:[%d]", rep.nul);
+        return SOCKS4_REPLY_KO;
     }
     
     printl(LOG_VERB, "SOCKS4 server reply status: %d", rep.status);
@@ -134,7 +134,7 @@ int socks5_hello(int socket, unsigned int auth_method, ...) {
     
     /* Veryfy Socks version */
     if (rep.ver != PROXY_PROTO_SOCKS_V5) {
-        printl(LOG_CRIT, "SOCKS5 server speaks unsupported protocol: v[%d]", rep.ver);
+        printl(LOG_CRIT, "SOCKS5 server unsupported protocol: v[%d]", rep.ver);
         return AUTH_METHOD_NOACCEPT;
     }
     
@@ -270,9 +270,9 @@ int socks5_request(int socket, uint8_t cmd, uint8_t atype, struct sockaddr *dadd
     printl(LOG_VERB, "SOCKS5 reply: [%d][%d], Bytes [%d]", rep->ver, 
         rep->status, rcount);
 
-    if (rep->ver != PROXY_PROTO_SOCKS_V5) {
-        printl(LOG_CRIT, "SOCKS5 server speaks unsupported protocol version: [%d]", rep->ver);
-        mexit(1, pfile_name);
+    if (rep->ver != PROXY_PROTO_SOCKS_V5) {                             /* Report SOCKS5 general failure */
+        printl(LOG_WARN, "SOCKS5 server speaks unsupported protocol v:[%d]", rep->ver);
+        return SOCKS5_REPLY_KO;
     }
     
     printl(LOG_VERB, "SOCKS5 server reply status: %d", rep->status);